WFLY-15165 Drop deprecated interfaces/classes/methods from wildfly-clustering-service
