--- clustering/service/src/main/java/org/wildfly/clustering/service/SuppliedValueService.java
@@ -1,50 +1,0 @@
-/*
- * JBoss, Home of Professional Open Source.
- * Copyright 2016, Red Hat, Inc., and individual contributors
- * as indicated by the @author tags. See the copyright.txt file in the
- * distribution for a full listing of individual contributors.
- *
- * This is free software; you can redistribute it and/or modify it
- * under the terms of the GNU Lesser General Public License as
- * published by the Free Software Foundation; either version 2.1 of
- * the License, or (at your option) any later version.
- *
- * This software is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public
- * License along with this software; if not, write to the Free
- * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
- * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
- */
-
-package org.wildfly.clustering.service;
-
-import java.util.function.Consumer;
-import java.util.function.Function;
-import java.util.function.Supplier;
-
-/**
- * Generic {@link org.jboss.msc.service.Service} whose provided value is a mapping of a supplied value.
- * @author Paul Ferraro
- * @deprecated Replaced by {@link FunctionalService}.
- */
-@Deprecated
-public class SuppliedValueService<T, V> extends FunctionalValueService<T, V> {
-
-    /**
-     * Constructs a new supplied value service.
-     * @param mapper a function that maps the supplied value to the service value
-     * @param supplier produces the supplied value
-     * @param destroyer a consumer that destroys the supplied value
-     */
-    public SuppliedValueService(Function<T, V> mapper, Supplier<T> supplier, Consumer<T> destroyer) {
-        super(mapper, context -> supplier.get(), (context, value) -> {
-            if (destroyer != null) {
-                destroyer.accept(value);
-            }
-        });
-    }
-}
