+++ clustering/singleton/extension/src/main/java/org/wildfly/extension/clustering/singleton/SingletonServiceHandler.java
@@ -33,7 +33,6 @@ import org.jboss.as.controller.PathAddress;
 import org.jboss.dmr.ModelNode;
 import org.jboss.msc.service.ServiceName;
 import org.jboss.msc.service.ServiceTarget;
-import org.wildfly.clustering.service.AliasServiceBuilder;
 import org.wildfly.clustering.service.IdentityServiceConfigurator;
 import org.wildfly.clustering.singleton.SingletonRequirement;
 import org.wildfly.extension.clustering.singleton.SingletonResourceDefinition.Capability;
